enh(embeddings): Move to quantized llama.cpp gguf model
