+++ config.gpu.yaml
@@ -27,20 +27,20 @@ vectordb:
     # url: http://localhost:8080
 
 embedding:
-  hugging_face:
-    model_name: nextcloud-ai/multilingual-e5-large-instruct
-    model_kwargs:
-      device: cuda
+  llama:
+    model_path: multilingual-e5-large-instruct-q6_k.gguf
+    n_batch: 512
+    n_ctx: 8192
 
   instructor:
     model_name: hkunlp/instructor-base
     model_kwargs:
-      device: cpu
+      device: cuda
 
-  llama:
-    model_path: dolphin-2.2.1-mistral-7b.Q5_K_M.gguf
-    n_batch: 512
-    n_ctx: 8192
+  hugging_face:
+    model_name: nextcloud-ai/multilingual-e5-large-instruct
+    model_kwargs:
+      device: cuda
 
 llm:
   nc_texttotext:
