+++ main.py
@@ -10,10 +10,8 @@ from context_chat_backend.utils import to_int  # isort: skip
 
 if __name__ == '__main__':
 	app_config = app.extra['CONFIG']
-	enabled = app.extra['ENABLED']
 
 	print('App config:\n' + dumps(app_config, indent=2), flush=True)
-	print('\n\nApp', 'enabled' if app.extra['ENABLED'] else 'disabled', 'at startup', flush=True)
 
 	uvicorn.run(
 		app='context_chat_backend:app',
